refactor: SPDV-435
