+++ src/modules/filemanager/components/FMInitialModal/FMInitialModal.tsx
@@ -1,11 +1,11 @@
 import { ReactElement, useContext, useEffect, useRef, useState } from 'react'
 
-import { BZZ, Duration, RedundancyLevel, Size, Utils } from '@ethersphere/bee-js'
+import { Duration, RedundancyLevel, Size, Utils } from '@ethersphere/bee-js'
 import CircularProgress from '@material-ui/core/CircularProgress'
 import './FMInitialModal.scss'
 import { CustomDropdown } from '../CustomDropdown/CustomDropdown'
 import { FMButton } from '../FMButton/FMButton'
-import { getExpiryDateByLifetime } from '../../utils/utils'
+import { fmFetchCost, getExpiryDateByLifetime } from '../../utils/utils'
 import { desiredLifetimeOptions } from '../../constants/constants'
 import { Context as SettingsContext } from '../../../../providers/Settings'
 import { FMSlider } from '../FMSlider/FMSlider'
@@ -63,36 +63,8 @@ export function FMInitialModal({ handleVisibility }: FMInitialModalProps): React
   }, [erasureCodeLevel])
 
   useEffect(() => {
-    const fetchCost = async () => {
-      if (currentFetch.current) {
-        await currentFetch.current
-      }
-      const fetchPromise = (async () => {
-        let cost: BZZ | undefined = undefined
-        try {
-          if (Size.fromBytes(capacity).toGigabytes() >= 0 && validityEndDate.getTime() >= new Date().getTime()) {
-            cost = await beeApi?.getStorageCost(
-              Size.fromBytes(capacity),
-              Duration.fromEndDate(validityEndDate),
-              undefined,
-              false,
-              erasureCodeLevel,
-            )
-            setCost(cost ? cost.toSignificantDigits(2) : '0')
-          } else {
-            setCost('0')
-          }
-        } catch (e) {
-          //TODO It needs to be discussed what happens to the error
-        }
-      })()
-      currentFetch.current = fetchPromise
-      await fetchPromise
-      currentFetch.current = null
-    }
-
     if (validityEndDate.getTime() > new Date().getTime()) {
-      fetchCost()
+      fmFetchCost(capacity, validityEndDate, false, erasureCodeLevel, beeApi, setCost, currentFetch)
 
       if (lifetimeIndex >= 0) {
         setIsCreateEnabled(true)
